10-7

+++ server/migrations/005-order.js
@@ -18,7 +18,7 @@ module.exports = {
         }
       },
       order_total: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
